pyproject.toml
[tool.poetry]
name = "django-app"
version = "0.1.0"
description = ""
authors = ["PiyushVyas13 <jaiprakashvyas6@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_app"}]

[tool.poetry.dependencies]
python = "^3.9.7"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
django-rest-knox = "^4.2.0"
django-cors-headers = "^3.14.0"
reportlab = "^3.6.12"
ipython = "^8.13.2"
pdfrw = "^0.4"
pypdf2 = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
